feat(api): add full step data to workflow dto; refactor
