+++ apps/dashboard/src/utils/types.ts
@@ -1,4 +1,4 @@
-import type { StepResponseDto } from '@novu/shared';
+import type { StepDataDto } from '@novu/shared';
 
 export enum ConnectionStatus {
   CONNECTED = 'connected',
@@ -20,6 +20,4 @@ export type RuntimeIssue = {
   message: string;
 };
 
-export type Step = Pick<StepResponseDto, 'name' | 'type' | '_id' | 'stepId' | 'issues'> & {
-  slug: string;
-};
+export type Step = Pick<StepDataDto, 'name' | 'type' | '_id' | 'stepId' | 'issues' | 'slug'>;
