+++ packages/shared/src/dto/workflows/get-list-query-params.ts
@@ -1,4 +1,4 @@
-import { WorkflowResponseDto } from './workflow-response.dto';
+import { WorkflowResponseDto } from './workflow.dto';
 import { LimitOffsetPaginationDto } from '../../types';
 
 export class GetListQueryParams extends LimitOffsetPaginationDto<WorkflowResponseDto, 'updatedAt'> {
