+++ packages/shared/src/dto/workflows/index.ts
@@ -6,9 +6,9 @@ export * from './json-schema-dto';
 export * from './preview-step-response.dto';
 export * from './promote-workflow-dto';
 export * from './step-content-issue.enum';
-export * from './step-data.dto';
+export * from './step.dto';
+export * from './workflow.dto';
 export * from './update-workflow-deprecated.dto';
-export * from './workflow-response.dto';
+export * from './workflow-deprecated.dto';
 export * from './workflow-status-enum';
 export * from './workflow-test-data-response-dto';
-export * from './workflow.dto';
