+++ packages/shared/src/dto/workflows/preview-step-response.dto.ts
@@ -1,6 +1,5 @@
 import { ActionTypeEnum, ChannelTypeEnum } from '../../types';
 import { SubscriberDto } from '../subscriber';
-import { ContentIssue } from './workflow-response.dto';
 
 export class RenderOutput {}
