+++ packages/shared/src/dto/workflows/workflow.dto.ts
@@ -1,33 +1,95 @@
-import { IWorkflowStepMetadata } from '../../entities/step';
-import { BuilderFieldType, BuilderGroupValues, FilterParts } from '../../types';
-import { MessageTemplateDto } from '../message-template';
-import { UpdateWorkflowDto, StepCreateDto, StepUpdateDto } from './workflow-response.dto';
+import type { JSONSchemaDto } from './json-schema-dto';
+import { Slug } from '../../types/utils';
+import { WorkflowCreationSourceEnum, StepTypeEnum, WorkflowOriginEnum, WorkflowPreferences } from '../../types';
+import { WorkflowStatusEnum } from './workflow-status-enum';
+import { StepCreateDto, StepUpdateDto, StepDataDto } from './step.dto';
 
-export class StepVariantDto {
-  id?: string;
-  _id?: string;
-  name?: string;
-  uuid?: string;
-  _templateId?: string;
-  template?: MessageTemplateDto;
-  filters?: {
-    isNegated?: boolean;
-    type?: BuilderFieldType;
-    value?: BuilderGroupValues;
-    children?: FilterParts[];
-  }[];
+export class ControlsSchema {
+  schema: JSONSchemaDto;
+}
+
+export type PatchWorkflowDto = {
   active?: boolean;
-  shouldStopOnFail?: boolean;
-  replyCallback?: {
-    active: boolean;
-    url?: string;
-  };
-  metadata?: IWorkflowStepMetadata;
+  name?: string;
+  description?: string;
+  tags?: string[];
+};
+
+export type ListWorkflowResponse = {
+  workflows: WorkflowListResponseDto[];
+  totalCount: number;
+};
+
+export type WorkflowListResponseDto = Pick<
+  WorkflowResponseDto,
+  'name' | 'tags' | 'updatedAt' | 'createdAt' | '_id' | 'workflowId' | 'slug' | 'status' | 'origin'
+> & {
+  stepTypeOverviews: StepTypeEnum[];
+};
+
+export type WorkflowCommonsFields = {
+  name: string;
+  description?: string;
+  tags?: string[];
+  active?: boolean;
+};
+
+export type PreferencesResponseDto = {
+  user: WorkflowPreferences | null;
+  default: WorkflowPreferences;
+};
+
+export type PreferencesRequestDto = {
+  user: WorkflowPreferences | null;
+  workflow?: WorkflowPreferences | null;
+};
+
+export type WorkflowResponseDto = WorkflowCommonsFields & {
+  _id: string;
+  workflowId: string;
+  slug: Slug;
+  updatedAt: string;
+  createdAt: string;
+  steps: StepDataDto[];
+  origin: WorkflowOriginEnum;
+  preferences: PreferencesResponseDto;
+  status: WorkflowStatusEnum;
+  issues?: Record<WorkflowCreateAndUpdateKeys, RuntimeIssue>;
+};
+export type WorkflowCreateAndUpdateKeys = keyof CreateWorkflowDto | keyof UpdateWorkflowDto;
+export class RuntimeIssue {
+  issueType: WorkflowIssueTypeEnum;
+  variableName?: string;
+  message: string;
+}
+export enum WorkflowIssueTypeEnum {
+  MISSING_VALUE = 'MISSING_VALUE',
+  MAX_LENGTH_ACCESSED = 'MAX_LENGTH_ACCESSED',
+  WORKFLOW_ID_ALREADY_EXISTS = 'WORKFLOW_ID_ALREADY_EXISTS',
+  DUPLICATED_VALUE = 'DUPLICATED_VALUE',
+  LIMIT_REACHED = 'LIMIT_REACHED',
 }
 
-export class NotificationStepDto extends StepVariantDto {
-  variants?: StepVariantDto[];
-}
+export type CreateWorkflowDto = WorkflowCommonsFields & {
+  workflowId: string;
+
+  steps: StepCreateDto[];
+
+  __source: WorkflowCreationSourceEnum;
+
+  preferences?: PreferencesRequestDto;
+};
+
+export type UpdateWorkflowDto = WorkflowCommonsFields & {
+  /**
+   * We allow to update workflow id only for code first workflows
+   */
+  workflowId?: string;
+
+  steps: (StepCreateDto | StepUpdateDto)[];
+
+  preferences: PreferencesRequestDto;
+};
 
 export type UpsertWorkflowBody = Omit<UpdateWorkflowDto, 'steps'> & {
   steps: UpsertStepBody[];
